+++ examples/example-remote/config/extraScript.js
@@ -1,5 +1,0 @@
-export const headScripts = [
-  {
-    src: '/interactionConfig/initLocal.core.js',
-  },
-];