+++ examples/example-remote/public/interactionConfig/initLocal.development.js
@@ -1,6 +1,0 @@
-window.appInitCustomLocalSpecial = {
-  // 控制台显示日志调试信息
-  showLogInConsole: true,
-  // local storage 数据加密
-  localStorageSecretSwitch: false,
-};