+++ examples/example-remote/src/components/PageLoading/index.jsx
@@ -1,5 +1,0 @@
-// loading components from code split
-// https://umijs.org/plugin/umi-plugin-react.html#dynamicimport
-
-// eslint-disable-next-line import/named
-export { PageLoading } from '@ant-design/pro-components';