+++ examples/example-remote/src/components/index.jsx
@@ -1,3 +1,0 @@
-export * from './FilePreviewDrawer';
-export * from './MenuCard';
-export * from './PageLoading';