+++ examples/example-remote/src/customConfig/accessWayCollection.js
@@ -1,7 +1,0 @@
-import { accessWayCustomCollection } from './custom/accessWayCollection';
-import { accessWayInfrastructureCollection } from './general/accessWayCollection';
-
-export const accessWayCollection = {
-  ...accessWayInfrastructureCollection,
-  ...accessWayCustomCollection,
-};