+++ examples/example-remote/src/customConfig/index.js
@@ -1,3 +1,0 @@
-export * from './accessWayCollection';
-export * from './custom';
-export * from './general';