+++ examples/example-remote/src/customSpecialComponents/CustomAssembly/index.js
@@ -1,3 +1,0 @@
-export * from './footer';
-export * from './menuHeader';
-export * from './timeAndOperator';