+++ examples/example-remote/src/customSpecialComponents/FunctionSpecialCustomComponent/index.jsx
@@ -1,2 +1,0 @@
-export * from './YonYouImportHistoryStatus';
-export * from './YonYouPushMessageStatus';