+++ examples/example-remote/src/customSpecialComponents/index.jsx
@@ -1,6 +1,0 @@
-export * from './BaseUpdateKeyValueInfoModal';
-export * from './CustomAssembly';
-export * from './DataTabContainerSupplement';
-export * from './FunctionSpecialCustomComponent';
-export * from './FunctionSpecialGeneralComponent';
-export * from './UpdateModuleModalBase';