+++ examples/example-remote/src/locales/zh-CN/menu.js
@@ -1,6 +1,0 @@
-import { menuZhCNCustom, menuZhCNGeneral } from '../../localeConfig';
-
-export default {
-  ...menuZhCNCustom,
-  ...menuZhCNGeneral,
-};