+++ examples/example-remote/src/pageBases/FlowCase/index.jsx
@@ -1,7 +1,0 @@
-export * from './Assist';
-export * from './BaseFlowCaseFormInfoDrawer';
-export * from './BaseFlowCasePageListLatestApproveDrawer';
-export * from './BaseFlowCasePageListUnderwayDrawer';
-export * from './BaseFlowCasePageListWaitApproveDrawer';
-export * from './BaseFlowCaseProcessChainDrawer';
-export * from './BaseFlowCaseUpdateBasicInfoDrawer';