+++ examples/example-remote/src/pageBases/FlowCaseNextProcessApprove/index.jsx
@@ -1,3 +1,0 @@
-export * from './BaseFlowCaseNextProcessApprovePageList';
-export * from './BaseFlowCaseNextProcessApprovePageListDrawer';
-export * from './BaseFlowCaseNextProcessApprovePreviewDrawer';