+++ examples/example-remote/src/pageBases/FlowCaseNextProcessNotification/index.jsx
@@ -1,3 +1,0 @@
-export * from './BaseFlowCaseNextProcessNotificationPageList';
-export * from './BaseFlowCaseNextProcessNotificationPageListDrawer';
-export * from './BaseFlowCaseNextProcessNotificationPreviewDrawer';