+++ examples/example-remote/src/pageBases/FlowCaseNextProcessProgress/index.jsx
@@ -1,3 +1,0 @@
-export * from './BaseFlowCaseNextProcessProgressPageList';
-export * from './BaseFlowCaseNextProcessProgressPageListDrawer';
-export * from './BaseFlowCaseNextProcessProgressPreviewDrawer';