+++ examples/example-remote/src/pageBases/FlowCaseProcessHistory/index.jsx
@@ -1,6 +1,0 @@
-export * from './Assist';
-export * from './BaseFlowCaseProcessHistoryPageList';
-export * from './BaseFlowCaseProcessHistoryPageListDrawer';
-export * from './BaseFlowCaseProcessHistoryPassModal';
-export * from './BaseFlowCaseProcessHistoryPreviewDrawer';
-export * from './BaseFlowCaseProcessHistoryRefuseModal';