+++ examples/example-remote/src/pageBases/FlowFormDesign/index.jsx
@@ -1,2 +1,0 @@
-export * from './Assist';
-export * from './BaseFlowCaseFormDocumentDrawer';