+++ examples/example-remote/src/pages/DevelopInfo/TabPageBase/index.jsx
@@ -1,7 +1,0 @@
-import { DataForm } from 'antd-management-fast-framework';
-
-const { BaseUpdateFormTab } = DataForm;
-
-class TabPageBase extends BaseUpdateFormTab {}
-
-export { TabPageBase };