+++ examples/example-remote/src/pages/SmsCategoryStatistic/Assist/tools.js
@@ -1,8 +1,0 @@
-export function supplementPlanSaleId(o, properties) {
-  const { externalData } = properties;
-  const { smsCategoryStatisticId } = externalData;
-
-  const result = { ...o, smsCategoryStatisticId: smsCategoryStatisticId || '' };
-
-  return result;
-}