+++ examples/example-remote/src/pages/WorkflowFormDesign/Assist/action.js
@@ -1,1 +1,0 @@
-export * from '../../../pageBases/FlowFormDesign/Assist/action';