+++ examples/example-remote/src/pages/WorkflowLine/Common/data.js
@@ -1,5 +1,0 @@
-import { fieldDataFlowLine } from '../../../customConfig';
-
-export const fieldData = {
-  ...fieldDataFlowLine,
-};