+++ examples/example-remote/src/pages/WorkflowNode/Common/data.js
@@ -1,5 +1,0 @@
-import { fieldDataFlowNode } from '../../../customConfig';
-
-export const fieldData = {
-  ...fieldDataFlowNode,
-};