+++ examples/example-remote/src/utils/index.jsx
@@ -1,6 +1,0 @@
-// 此文件将会从模板库自动更新，请勿改动此文件内容。
-
-export * from './custom';
-export * from './general';
-export * from './init';
-export * from './theme';