+++ examples/example-remote/test/simple.test.js
@@ -1,5 +1,0 @@
-describe('group test description', () => {
-  test('simple test will be true', () => {
-    expect(true).toBe(true);
-  });
-});