+++ packages/antd-management-fast-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 2.6.158
+
+### Patch Changes
+
+- update
+
+## 2.6.157
+
+### Patch Changes
+
+- update
+
 ## 2.6.156
 
 ### Patch Changes