+++ packages/antd-management-fast-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.1.135
+
+### Patch Changes
+
+- update
+
+## 1.1.134
+
+### Patch Changes
+
+- update
+
 ## 1.1.133
 
 ### Patch Changes