+++ packages/antd-management-fast-develop/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.1.146
+
+### Patch Changes
+
+- update
+
+## 1.1.145
+
+### Patch Changes
+
+- update
+
 ## 1.1.144
 
 ### Patch Changes